requirements.txt
fastapi
numpy
uvicorn
tensorflow
keras
pandas
python-multipart
Pillow